change-comments: Add publication existence verification and remove access to other repositories from comment repository

+++ src/Interaction/Comment/Application/DeleteCommentUseCase.py
@@ -1,13 +1,20 @@
 import inject
 from src.Interaction.Comment.Domain.CommentRepository import CommentRepository
+from src.Publication.Application.VerifyPublicationExistenceUseCase import (
+    VerifyPublicationExistenceUseCase,
+)
 
 
 class DeleteCommentUseCase:
     @inject.autoparams()
     def __init__(self, comment_repository: CommentRepository):
         self.comment_repository = comment_repository
+        self.verify_pub_existence_use_case = VerifyPublicationExistenceUseCase()
 
-    def execute(self, pub_id: str, comment_id: str, is_adoption: bool) -> None:
-        return self.comment_repository.delete_comment(
-            pub_id=pub_id, comment_id=comment_id, is_adoption=is_adoption
-        )
+    def execute(self, pub_id: str, comment_id: str) -> None:
+        if self.verify_pub_existence_use_case.execute(pub_id):
+            return self.comment_repository.delete_comment(
+                pub_id=pub_id, comment_id=comment_id
+            )
+        else:
+            raise Exception("Publication not found")
